Added normal photo key. Home function is working. 0.5.1.
The drone can now take normal photos as well, without plotting any
temperature values and saving them on the same directory.

The home function has been tested and it's working. Please, be careful
when using this drone as it will go high (as high as it can) and will
return home at max speed. It can be a bit scary, so please be aware
of this function and use it only when needed.

The speed of the drone has been increased, but it's still not the max
speed the drone can reach. This function will be available soon.